refactor: decompose the remote_account_provider impl monolith
Pure move: the 2,600-line generic RemoteAccountProvider impl is split
into provider_setup (construction, endpoint selection, transport
policy), provider_updates (update listening, slot-matched resolution,
replay outbox), provider_subscriptions (tiers, acquire/ensure/release,
program subs), and provider_fetch (RPC fetching, dedup, retries). Moved
methods gain pub(super); no logic changes. mod.rs retains shared types,
guards, and the concrete-client ctor (4,253 -> 1,563 lines).
