+++ magicblock-chainlink/src/remote_account_provider/mod.rs
@@ -1,6 +1,5 @@
 use std::{
     collections::{hash_map::Entry, HashMap, HashSet},
-    num::NonZeroUsize,
     sync::{
         atomic::{AtomicU64, Ordering},
         Arc, Mutex, RwLock, Weak,
@@ -17,32 +16,23 @@ use config::RemoteAccountProviderConfig;
 pub(crate) use errors::{
     RemoteAccountProviderError, RemoteAccountProviderResult,
 };
-use futures_util::future::{join_all, try_join_all};
-use lru_cache::TieredSubscribedAccountsTracker;
 pub use lru_cache::{AccountsLruCache, AddAccountOutcome};
-use magicblock_config::config::{GrpcConfig, SubscriptionTransport};
+use magicblock_config::config::GrpcConfig;
 pub(crate) use remote_account::RemoteAccount;
 pub use remote_account::RemoteAccountUpdateSource;
-use solana_account::Account;
-use solana_account_decoder_client_types::{
-    UiAccountEncoding, UiDataSliceConfig,
-};
 use solana_commitment_config::CommitmentConfig;
 use solana_pubkey::Pubkey;
-use solana_rpc_client_api::{
-    client_error::ErrorKind,
-    config::{RpcAccountInfoConfig, RpcProgramAccountsConfig},
-    custom_error::JSON_RPC_SERVER_ERROR_MIN_CONTEXT_SLOT_NOT_REACHED,
-    request::RpcError,
-};
-use solana_sdk_ids::sysvar::clock;
 use tokio::{
     sync::{mpsc, oneshot, Mutex as AsyncMutex, Notify},
-    task, time,
+    task,
 };
 use tracing::*;
 
 pub mod chain_slot;
+mod provider_fetch;
+mod provider_setup;
+mod provider_subscriptions;
+mod provider_updates;
 use chain_slot::ChainSlot;
 pub(crate) mod chain_laser_actor;
 pub mod chain_laser_client;
@@ -65,31 +55,18 @@ pub(crate) mod subscription_reconciler;
 mod tests;
 
 pub use endpoint::{Endpoint, Endpoints};
-use magicblock_metrics::{
-    metrics,
-    metrics::{
-        dec_chainlink_pending_fetch_waiters_gauge, inc_account_fetches_failed,
-        inc_account_fetches_found_with_context,
-        inc_account_fetches_not_found_with_context,
-        inc_account_fetches_success,
-        inc_chainlink_empty_placeholder_accounts_total_with_context,
-        inc_chainlink_pending_fetch_accounts_with_context,
-        inc_chainlink_pending_fetch_waiters_gauge,
-        inc_chainlink_pending_fetch_waiters_with_context,
-        inc_chainlink_subscription_cleanup_accounts,
-        inc_chainlink_subscription_registration_accounts,
-        inc_chainlink_subscription_release_accounts,
-        observe_chainlink_companion_fetch_attempts,
-        observe_chainlink_companion_fetch_duration_seconds,
-        observe_chainlink_pending_fetch_owner_duration_seconds_with_context,
-        set_monitored_accounts_count, AccountFetchContext, AccountFetchReason,
-        ChainlinkCompanionFetchKind, ChainlinkCompanionFetchOutcome,
-        ChainlinkEmptyPlaceholderStage, ChainlinkPendingFetchLayer,
-        ChainlinkPendingFetchOutcome, Outcome, SubscriptionCleanupOutcome,
-        SubscriptionCleanupSource, SubscriptionReasonLabel,
-        SubscriptionRegistrationOrigin, SubscriptionRegistrationOutcome,
-        SubscriptionReleaseOutcome,
-    },
+use magicblock_metrics::metrics::{
+    dec_chainlink_pending_fetch_waiters_gauge,
+    inc_chainlink_subscription_cleanup_accounts,
+    inc_chainlink_subscription_registration_accounts,
+    observe_chainlink_companion_fetch_attempts,
+    observe_chainlink_companion_fetch_duration_seconds,
+    observe_chainlink_pending_fetch_owner_duration_seconds_with_context,
+    AccountFetchContext, ChainlinkCompanionFetchKind,
+    ChainlinkCompanionFetchOutcome, ChainlinkPendingFetchLayer,
+    ChainlinkPendingFetchOutcome, SubscriptionCleanupOutcome,
+    SubscriptionCleanupSource, SubscriptionReasonLabel,
+    SubscriptionRegistrationOrigin, SubscriptionRegistrationOutcome,
 };
 pub use remote_account::{ResolvedAccount, ResolvedAccountSharedData};
 
@@ -1477,2617 +1454,6 @@ impl<T: ChainRpcClient, U: ChainPubsubClient> RemoteAccountProvider<T, U> {
         self.next_fetching_account_generation
             .fetch_add(1, Ordering::Relaxed)
             .wrapping_add(1)
-    }
-
-    pub async fn try_from_clients_and_mode(
-        rpc_client: T,
-        pubsub_client: U,
-        subscription_forwarder: mpsc::Sender<ForwardedSubscriptionUpdate>,
-        config: &RemoteAccountProviderConfig,
-        lrucache_subscribed_accounts: Arc<AccountsLruCache>,
-        chain_slot: Arc<AtomicU64>,
-    ) -> ChainlinkResult<Option<RemoteAccountProvider<T, U>>> {
-        let chain_slot = ChainSlot::new(chain_slot);
-        if config.lifecycle_mode().needs_remote_account_provider() {
-            Ok(Some(
-                Self::new(
-                    rpc_client,
-                    pubsub_client,
-                    subscription_forwarder,
-                    config,
-                    lrucache_subscribed_accounts,
-                    chain_slot,
-                )
-                .await?,
-            ))
-        } else {
-            Ok(None)
-        }
-    }
-
-    /// Creates a background task that periodically reconciles subscriptions
-    /// with the LRU (repairing missing ones, e.g. after a partial
-    /// resubscription) and optionally updates the active subscriptions gauge
-    #[allow(clippy::too_many_arguments)]
-    fn start_active_subscriptions_updater<PubsubClient: ChainPubsubClient>(
-        subscribed_accounts: Arc<AccountsLruCache>,
-        secondary_subscriptions: Arc<AccountsLruCache>,
-        confirmed_missing_subscriptions: Arc<Mutex<HashSet<Pubkey>>>,
-        pubsub_client: Arc<PubsubClient>,
-        removed_account_tx: mpsc::Sender<Pubkey>,
-        subscription_key_locks: SubscriptionKeyLocks,
-        subscription_ownership: SubscriptionOwnershipMap,
-        fetching_accounts: Arc<FetchingAccounts>,
-        capacity_eviction_protection: SharedCapacityEvictionProtectionPredicate,
-        emit_metrics: bool,
-    ) -> task::JoinHandle<()> {
-        task::spawn(async move {
-            let mut interval = time::interval(Duration::from_millis(
-                ACTIVE_SUBSCRIPTIONS_UPDATE_INTERVAL_MS,
-            ));
-            let never_evicted = subscribed_accounts.never_evicted_accounts();
-
-            loop {
-                interval.tick().await;
-                let pubsub_total =
-                    subscription_reconciler::reconcile_subscriptions(
-                        &subscribed_accounts,
-                        &secondary_subscriptions,
-                        &confirmed_missing_subscriptions,
-                        pubsub_client.as_ref(),
-                        &never_evicted,
-                        &removed_account_tx,
-                        Some(&subscription_key_locks),
-                        Some(&subscription_ownership),
-                        Some(fetching_accounts.as_ref()),
-                        Some(&capacity_eviction_protection),
-                    )
-                    .await;
-
-                debug!(count = pubsub_total, "Updating active subscriptions");
-                if emit_metrics {
-                    set_monitored_accounts_count(pubsub_total);
-                }
-            }
-        })
-    }
-
-    /// Creates a new instance of the remote account provider
-    /// By the time this method returns the current chain slot was resolved and
-    /// a subscription setup to keep it up to date.
-    pub(crate) async fn new(
-        rpc_client: T,
-        pubsub_client: U,
-        subscription_forwarder: mpsc::Sender<ForwardedSubscriptionUpdate>,
-        config: &RemoteAccountProviderConfig,
-        lrucache_subscribed_accounts: Arc<AccountsLruCache>,
-        chain_slot: ChainSlot,
-    ) -> RemoteAccountProviderResult<Self> {
-        let secondary_subscriptions = Arc::new(AccountsLruCache::new(
-            // SAFETY: config guarantees a non-zero capacity
-            NonZeroUsize::new(config.secondary_subscriptions_lru_capacity())
-                .expect("lru capacity must be non-zero"),
-        ));
-        Self::new_with_secondary_subscriptions(
-            rpc_client,
-            pubsub_client,
-            subscription_forwarder,
-            config,
-            lrucache_subscribed_accounts,
-            secondary_subscriptions,
-            chain_slot,
-        )
-        .await
-    }
-
-    async fn new_with_secondary_subscriptions(
-        rpc_client: T,
-        pubsub_client: U,
-        subscription_forwarder: mpsc::Sender<ForwardedSubscriptionUpdate>,
-        config: &RemoteAccountProviderConfig,
-        lrucache_subscribed_accounts: Arc<AccountsLruCache>,
-        secondary_subscriptions: Arc<AccountsLruCache>,
-        chain_slot: ChainSlot,
-    ) -> RemoteAccountProviderResult<Self> {
-        let (removed_account_tx, removed_account_rx) =
-            tokio::sync::mpsc::channel(100);
-        let subscription_key_locks: SubscriptionKeyLocks =
-            Arc::new(AsyncMutex::new(HashMap::new()));
-        let confirmed_missing_subscriptions =
-            Arc::new(Mutex::new(HashSet::new()));
-        let subscription_ownership: SubscriptionOwnershipMap =
-            Arc::new(AsyncMutex::new(HashMap::new()));
-        let fetching_accounts = Arc::<FetchingAccounts>::default();
-        let capacity_eviction_protection:
-            SharedCapacityEvictionProtectionPredicate =
-            Arc::new(RwLock::new(None));
-
-        // The reconciler always runs: partial resubscriptions rely on it for
-        // repair. The config flag only gates the metrics emission.
-        let active_subscriptions_updater =
-            Some(Self::start_active_subscriptions_updater(
-                lrucache_subscribed_accounts.clone(),
-                secondary_subscriptions.clone(),
-                confirmed_missing_subscriptions.clone(),
-                Arc::new(pubsub_client.clone()),
-                removed_account_tx.clone(),
-                subscription_key_locks.clone(),
-                subscription_ownership.clone(),
-                fetching_accounts.clone(),
-                capacity_eviction_protection.clone(),
-                config.enable_subscription_metrics(),
-            ));
-
-        let me = Self {
-            fetching_accounts,
-            next_fetching_account_generation: AtomicU64::default(),
-            subscription_ownership,
-            subscription_transition_lock: Arc::new(AsyncMutex::new(())),
-            subscription_key_locks,
-            rpc_client,
-            pubsub_client,
-            chain_slot,
-            last_update_slot: Arc::<AtomicU64>::default(),
-            received_updates_count: Arc::<AtomicU64>::default(),
-            lrucache_subscribed_accounts,
-            secondary_subscriptions,
-            confirmed_missing_subscriptions,
-            capacity_eviction_protection,
-            subscription_forwarder: Arc::new(subscription_forwarder),
-            replay_outbox: Arc::default(),
-            replay_notify: Arc::new(Notify::new()),
-            removed_account_tx,
-            removed_account_rx: Mutex::new(Some(removed_account_rx)),
-            _active_subscriptions_task_handle: active_subscriptions_updater,
-        };
-
-        let updates = me.pubsub_client.take_updates();
-        me.listen_for_account_updates(updates)?;
-        me.start_replay_outbox_worker();
-        let clock_remote_account = me
-            .try_get(
-                clock::ID,
-                AccountFetchContext::internal(AccountFetchReason::Clock),
-            )
-            .await?;
-        match clock_remote_account {
-            RemoteAccount::NotFound(_) => {
-                Err(RemoteAccountProviderError::ClockAccountCouldNotBeResolved(
-                    clock::ID.to_string(),
-                ))
-            }
-            RemoteAccount::Found(_) => {
-                me.chain_slot.update(clock_remote_account.slot());
-                Ok(me)
-            }
-        }
-    }
-
-    pub async fn try_new_from_endpoints(
-        endpoints: &Endpoints,
-        commitment: CommitmentConfig,
-        subscription_forwarder: mpsc::Sender<ForwardedSubscriptionUpdate>,
-        config: &RemoteAccountProviderConfig,
-        chain_slot: Arc<AtomicU64>,
-    ) -> RemoteAccountProviderResult<
-        RemoteAccountProvider<
-            ChainRpcClientImpl,
-            SubMuxClient<ChainUpdatesClient>,
-        >,
-    > {
-        if endpoints.is_empty() {
-            return Err(
-                RemoteAccountProviderError::AccountSubscriptionsTaskFailed(
-                    "No endpoints provided".to_string(),
-                ),
-            );
-        }
-
-        // Build RPC clients (use the first RPC endpoint found)
-        let rpc_url = endpoints.rpc_url().ok_or_else(|| {
-            RemoteAccountProviderError::AccountSubscriptionsTaskFailed(
-                "No RPC endpoint found".to_string(),
-            )
-        })?;
-        let rpc_client =
-            ChainRpcClientImpl::new_from_url(rpc_url.as_str(), commitment);
-
-        // Single-transport subscriptions: every client speaks the configured
-        // transport. gRPC is the only supported production transport; with
-        // it configured but unavailable, the validator logs an error and
-        // runs degraded on websockets rather than refusing to start.
-        let (grpc, ws): (Vec<_>, Vec<_>) = endpoints
-            .pubsubs()
-            .into_iter()
-            .cloned()
-            .partition(|ep| matches!(ep, Endpoint::Grpc { .. }));
-        let grpc_selected = !grpc.is_empty();
-        let selected = match config.subscription_transport() {
-            SubscriptionTransport::Grpc if !grpc.is_empty() => {
-                metrics::set_subscription_transport_grpc(true);
-                grpc
-            }
-            SubscriptionTransport::Grpc => {
-                error!(
-                    "subscription-transport is grpc but no gRPC remote is \
-                     configured; running DEGRADED on websockets — this is \
-                     unsupported in production"
-                );
-                metrics::set_subscription_transport_grpc(false);
-                ws.clone()
-            }
-            SubscriptionTransport::Ws => {
-                warn!(
-                    "subscription-transport is ws: dev/test only, \
-                     unsupported in production"
-                );
-                metrics::set_subscription_transport_grpc(false);
-                ws.clone()
-            }
-        };
-
-        match Self::try_new_from_pubsubs(
-            selected,
-            commitment,
-            rpc_client.clone(),
-            chain_slot.clone(),
-            subscription_forwarder.clone(),
-            config,
-        )
-        .await
-        {
-            Ok(provider) => Ok(provider),
-            Err(err)
-                if matches!(
-                    config.subscription_transport(),
-                    SubscriptionTransport::Grpc
-                ) && grpc_selected
-                    && !ws.is_empty() =>
-            {
-                error!(
-                    error = %err,
-                    "gRPC subscriptions failed at startup; running DEGRADED \
-                     on websockets — this is unsupported in production"
-                );
-                metrics::set_subscription_transport_grpc(false);
-                Self::try_new_from_pubsubs(
-                    ws,
-                    commitment,
-                    rpc_client,
-                    chain_slot,
-                    subscription_forwarder,
-                    config,
-                )
-                .await
-            }
-            Err(err) => Err(err),
-        }
-    }
-
-    async fn try_new_from_pubsubs(
-        pubsub_endpoints: Vec<Endpoint>,
-        commitment: CommitmentConfig,
-        rpc_client: ChainRpcClientImpl,
-        chain_slot: Arc<AtomicU64>,
-        subscription_forwarder: mpsc::Sender<ForwardedSubscriptionUpdate>,
-        config: &RemoteAccountProviderConfig,
-    ) -> RemoteAccountProviderResult<
-        RemoteAccountProvider<
-            ChainRpcClientImpl,
-            SubMuxClient<ChainUpdatesClient>,
-        >,
-    > {
-        let resubscription_delay = config.resubscription_delay();
-        let pubsub_futs = pubsub_endpoints.into_iter().map(|ep| {
-            connect_pubsub_client(
-                ep,
-                commitment,
-                rpc_client.clone(),
-                chain_slot.clone(),
-                resubscription_delay,
-                config.grpc().clone(),
-            )
-        });
-        let results = join_all(pubsub_futs).await;
-        let pubsubs = collect_connected_pubsubs(results);
-
-        if pubsubs.is_empty() {
-            return Err(RemoteAccountProviderError::AllPubsubClientsFailed);
-        }
-        let subscribed_accounts = Arc::new(AccountsLruCache::new({
-            // SAFETY: NonZeroUsize::new only returns None if the value is 0.
-            // RemoteAccountProviderConfig can only be constructed with
-            // capacity > 0
-            let cap = config.subscribed_accounts_lru_capacity();
-            NonZeroUsize::new(cap).expect("non-zero capacity")
-        }));
-        let secondary_subscriptions = Arc::new(AccountsLruCache::new({
-            let cap = config.secondary_subscriptions_lru_capacity();
-            NonZeroUsize::new(cap).expect("non-zero capacity")
-        }));
-        let subscribed_accounts_tracker =
-            Arc::new(TieredSubscribedAccountsTracker::new(
-                subscribed_accounts.clone(),
-                secondary_subscriptions.clone(),
-            ));
-
-        let submux =
-            SubMuxClient::new(pubsubs, subscribed_accounts_tracker, None);
-
-        if !config.program_subs().is_empty() {
-            let count = config.program_subs().len();
-            debug!(count, "Subscribing to program accounts");
-            let subscribe_program_futs = config
-                .program_subs()
-                .iter()
-                .map(|program_id| submux.subscribe_program(*program_id));
-            try_join_all(subscribe_program_futs).await?;
-        }
-
-        let provider = RemoteAccountProvider::<
-            ChainRpcClientImpl,
-            SubMuxClient<ChainUpdatesClient>,
-        >::new_with_secondary_subscriptions(
-            rpc_client,
-            submux,
-            subscription_forwarder,
-            config,
-            subscribed_accounts,
-            secondary_subscriptions,
-            ChainSlot::new(chain_slot),
-        )
-        .await?;
-        Ok(provider)
-    }
-
-    pub(crate) fn promote_accounts(&self, pubkeys: &[&Pubkey]) {
-        self.lrucache_subscribed_accounts.promote_multi(pubkeys);
-        // This runs on the per-transaction ensure path; the secondary tier
-        // only holds fetch-owned/missing accounts and is empty in the common
-        // case, so skip its lock entirely then. A promote missed due to a
-        // concurrent insert is harmless (LRU ordering is a heuristic).
-        if !self.secondary_subscriptions.is_vacant() {
-            self.secondary_subscriptions.promote_multi(pubkeys);
-        }
-    }
-
-    pub(crate) async fn get_slot(&self) -> RemoteAccountProviderResult<u64> {
-        tokio::time::timeout(RPC_FETCH_TIMEOUT, self.rpc_client.get_slot())
-            .await
-            .map_err(|_| {
-                RemoteAccountProviderError::AccountResolutionsFailed(format!(
-                    "RPC call timeout fetching slot after {}ms",
-                    RPC_FETCH_TIMEOUT.as_millis()
-                ))
-            })?
-            .map_err(|err| {
-                RemoteAccountProviderError::AccountResolutionsFailed(format!(
-                    "RpcError fetching slot: {err:?}"
-                ))
-            })
-    }
-
-    pub(crate) async fn get_program_accounts_with_config(
-        &self,
-        pubkey: &Pubkey,
-        mut config: RpcProgramAccountsConfig,
-    ) -> RemoteAccountProviderResult<Vec<(Pubkey, Account)>> {
-        config.account_config.commitment = Some(self.rpc_client.commitment());
-
-        tokio::time::timeout(RPC_FETCH_TIMEOUT, async {
-            self.rpc_client
-                .get_program_accounts_with_config(pubkey, config)
-                .await
-        })
-        .await
-        .map_err(|_| {
-            RemoteAccountProviderError::AccountResolutionsFailed(format!(
-                "RPC call timeout fetching program accounts for {pubkey} after {}ms",
-                RPC_FETCH_TIMEOUT.as_millis()
-            ))
-        })?
-        .map_err(|err| {
-            RemoteAccountProviderError::AccountResolutionsFailed(format!(
-                "RpcError fetching program accounts for {pubkey}: {err:?}"
-            ))
-        })
-    }
-
-    pub(crate) fn set_capacity_eviction_protection<F>(&self, predicate: F)
-    where
-        F: Fn(&Pubkey) -> CapacityEvictionProtection + Send + Sync + 'static,
-    {
-        let mut guard = self
-            .capacity_eviction_protection
-            .write()
-            .unwrap_or_else(|poison| poison.into_inner());
-        *guard = Some(Arc::new(predicate));
-    }
-
-    pub fn try_get_removed_account_rx(
-        &self,
-    ) -> RemoteAccountProviderResult<mpsc::Receiver<Pubkey>> {
-        let mut rx = self
-            .removed_account_rx
-            .lock()
-            .expect("removed_account_rx lock poisoned");
-        rx.take().ok_or_else(|| {
-            RemoteAccountProviderError::LruCacheRemoveAccountSenderSupportsSingleReceiverOnly
-        })
-    }
-
-    pub fn chain_slot(&self) -> u64 {
-        self.chain_slot.load()
-    }
-
-    pub fn last_update_slot(&self) -> u64 {
-        self.last_update_slot.load(Ordering::Relaxed)
-    }
-
-    pub fn received_updates_count(&self) -> u64 {
-        self.received_updates_count.load(Ordering::Relaxed)
-    }
-
-    fn listen_for_account_updates(
-        &self,
-        mut updates: mpsc::Receiver<SubscriptionUpdate>,
-    ) -> RemoteAccountProviderResult<()> {
-        let fetching_accounts = self.fetching_accounts.clone();
-        let chain_slot = self.chain_slot.clone();
-        let received_updates_count = self.received_updates_count.clone();
-        let last_update_slot = self.last_update_slot.clone();
-        let subscription_forwarder = self.subscription_forwarder.clone();
-        let subscription_tiers = Arc::new(self.subscription_tier_ctx());
-        task::spawn(async move {
-            while let Some(update) = updates.recv().await {
-                let slot = update.slot;
-
-                received_updates_count.fetch_add(1, Ordering::Relaxed);
-                last_update_slot.store(slot, Ordering::Relaxed);
-
-                if update.pubkey == clock::ID {
-                    // We show as part of test_chain_pubsub_client_clock that the response
-                    // context slot always matches the slot encoded in the slot data.
-                    // Use fetch_max to ensure we always keep the highest slot value,
-                    // since GRPC may have already updated chain_slot to a higher value.
-                    chain_slot.update(slot);
-                    // NOTE: we do not forward clock updates
-                } else {
-                    trace!(
-                        pubkey = %update.pubkey,
-                        slot,
-                        "Received account update"
-                    );
-                    let remote_account = match update.account {
-                        Some(account) => RemoteAccount::from_fresh_account(
-                            account,
-                            slot,
-                            RemoteAccountUpdateSource::Subscription,
-                        ),
-                        None => {
-                            warn!(
-                                pubkey = %update.pubkey,
-                                "Account update could not be decoded"
-                            );
-                            RemoteAccount::NotFound(slot)
-                        }
-                    };
-
-                    let account_is_found = remote_account.is_found();
-
-                    // Fast path: fetch arbitration and tier movement only
-                    // apply while a fetch is pending or the account sits in
-                    // the secondary tier. All other updates forward without
-                    // taking the per-key guard or the transition lock.
-                    let needs_tier_handling =
-                        subscription_tiers.secondary.contains(&update.pubkey)
-                            || fetching_accounts
-                                .lock()
-                                .unwrap_or_else(|poison| poison.into_inner())
-                                .contains_key(&update.pubkey);
-
-                    // Serialize fetch arbitration and tier movement so a late
-                    // RPC result cannot overwrite this subscription update.
-                    let mut classification_error = None;
-                    let (forward_update, _accepted_update, resolved_fetch) =
-                        if !needs_tier_handling {
-                            // Record so a lagging RPC result cannot later win
-                            // classification against this newer update.
-                            if account_is_found {
-                                subscription_tiers
-                                    .record_classification(
-                                        update.pubkey,
-                                        slot,
-                                        SubscriptionClassificationSource::Subscription,
-                                    )
-                                    .await;
-                            }
-                            (
-                                Some(ForwardedSubscriptionUpdate {
-                                    pubkey: update.pubkey,
-                                    account: remote_account.clone(),
-                                    source: update.source,
-                                }),
-                                true,
-                                None,
-                            )
-                        } else {
-                            // The per-key guard serializes this update
-                            // against fetch resolutions and other transitions
-                            // of the same key; the tier helpers scope the
-                            // transition lock to their in-memory critical
-                            // sections.
-                            let _subscription_guard =
-                                subscription_key_owned_guard_from_map(
-                                    &subscription_tiers.subscription_key_locks,
-                                    update.pubkey,
-                                )
-                                .await;
-                            let classification_is_current = subscription_tiers
-                                .classification_is_current(
-                                    update.pubkey,
-                                    slot,
-                                    SubscriptionClassificationSource::Subscription,
-                                )
-                                .await;
-                            let result = if classification_is_current {
-                                let mut fetching =
-                                    fetching_accounts.lock().unwrap_or_else(
-                                        |poison| poison.into_inner(),
-                                    );
-                                if let Some(generation) = fetching
-                                    .get(&update.pubkey)
-                                    .map(|state| state.generation)
-                                {
-                                    if let Some(state) =
-                                    remove_fetching_account_if_generation_matches(
-                                        &mut fetching,
-                                        &update.pubkey,
-                                        generation,
-                                    )
-                                {
-                                    // If subscription update is newer than when we started fetching,
-                                    // resolve with the subscription data instead
-                                    if slot >= state.fetch_start_slot {
-                                        trace!(pubkey = %update.pubkey, slot = slot, fetch_start_slot = state.fetch_start_slot, generation, "Using subscription update instead of fetch");
-                                        metrics::observe_chainlink_pending_fetch_owner_duration_seconds_with_context(
-                                            state.fetch_context.clone(),
-                                            ChainlinkPendingFetchLayer::RemoteAccountProvider,
-                                            ChainlinkPendingFetchOutcome::ResolvedBySubscriptionUpdate,
-                                            state.owner_started_at.elapsed().as_secs_f64(),
-                                        );
-                                        metrics::inc_chainlink_pending_fetch_accounts_with_context(
-                                            state.fetch_context.clone(),
-                                            ChainlinkPendingFetchLayer::RemoteAccountProvider,
-                                            ChainlinkPendingFetchOutcome::ResolvedBySubscriptionUpdate,
-                                            1,
-                                        );
-
-                                        // Also forward: the fetch waiters may
-                                        // not clone the result (e.g. status
-                                        // reads) and dedup already dropped
-                                        // every other copy of this update.
-                                        (
-                                            Some(ForwardedSubscriptionUpdate {
-                                                pubkey: update.pubkey,
-                                                account: remote_account.clone(),
-                                                source: update.source,
-                                            }),
-                                            true,
-                                            Some((generation, state.waiters)),
-                                        )
-                                    } else {
-                                        // Subscription is stale, put the fetch tracking back
-                                        debug!(pubkey = %update.pubkey, slot = slot, fetch_start_slot = state.fetch_start_slot, generation, "Received stale subscription update");
-                                        fetching.insert(update.pubkey, state);
-                                        (None, false, None)
-                                    }
-                                } else {
-                                    (None, false, None)
-                                }
-                                } else {
-                                    (
-                                        Some(ForwardedSubscriptionUpdate {
-                                            pubkey: update.pubkey,
-                                            account: remote_account.clone(),
-                                            source: update.source,
-                                        }),
-                                        true,
-                                        None,
-                                    )
-                                }
-                            } else {
-                                debug!(pubkey = %update.pubkey, slot, "Ignoring stale subscription classification");
-                                (None, false, None)
-                            };
-
-                            // The in-flight acquisition may not have created
-                            // the ownership entry yet; record so the later
-                            // RPC result loses arbitration.
-                            let apply_classification = result.1
-                                && account_is_found
-                                && match result.2.as_ref() {
-                                    Some((generation, _)) => {
-                                        subscription_tiers
-                                            .record_classification_for_pending_fetch(
-                                                update.pubkey,
-                                                slot,
-                                                SubscriptionClassificationSource::Subscription,
-                                                *generation,
-                                            )
-                                            .await
-                                    }
-                                    None => {
-                                        subscription_tiers
-                                            .record_classification(
-                                                update.pubkey,
-                                                slot,
-                                                SubscriptionClassificationSource::Subscription,
-                                            )
-                                            .await
-                                    }
-                                };
-                            if apply_classification
-                                && !subscription_tiers
-                                    .secondary
-                                    .contains(&update.pubkey)
-                                && result.2.is_some()
-                                && !subscription_tiers
-                                    .primary
-                                    .contains(&update.pubkey)
-                            {
-                                // The pending fetch resolved before its
-                                // subscription setup created any tier state;
-                                // admit the found account into the primary
-                                // tier now so it is never handed to waiters
-                                // without primary admission.
-                                if let Err(err) = subscription_tiers
-                                    .admit_resolved_fetch_to_primary(
-                                        update.pubkey,
-                                    )
-                                    .await
-                                {
-                                    warn!(pubkey = %update.pubkey, error = ?err, "Failed to admit resolved-fetch account to primary subscription tier");
-                                    subscription_tiers
-                                        .clear_rejected_fetch_classification(
-                                            &update.pubkey,
-                                        )
-                                        .await;
-                                    classification_error =
-                                        Some(err.to_string());
-                                }
-                            } else if apply_classification
-                                && subscription_tiers
-                                    .secondary
-                                    .contains(&update.pubkey)
-                            {
-                                match subscription_tiers
-                                    .try_promote_found_to_primary(
-                                        update.pubkey,
-                                        true,
-                                    )
-                                    .await
-                                {
-                                    Ok(PromotionOutcome::Promoted) => {}
-                                    // The key was promoted by another
-                                    // transition while this one was in
-                                    // flight; it holds primary membership.
-                                    Ok(PromotionOutcome::NotInSecondary) => {}
-                                    // Evicted mid-promotion: the detached
-                                    // eviction cleanup owns the follow-up;
-                                    // the found update must not be forwarded
-                                    // without primary membership.
-                                    Ok(PromotionOutcome::Evicted) => {
-                                        classification_error = Some(
-                                            RemoteAccountProviderError::NoEvictableSubscriptionCapacity {
-                                                pubkey: update.pubkey,
-                                            }
-                                            .to_string(),
-                                        );
-                                    }
-                                    Ok(PromotionOutcome::NoCapacity) => {
-                                        let err = RemoteAccountProviderError::NoEvictableSubscriptionCapacity {
-                                            pubkey: update.pubkey,
-                                        };
-                                        subscription_tiers
-                                            .finalize_rejected_promotion(
-                                                &update.pubkey,
-                                            )
-                                            .await;
-                                        classification_error =
-                                            Some(err.to_string());
-                                    }
-                                    Err(err) => {
-                                        warn!(pubkey = %update.pubkey, error = ?err, "Failed to promote found account to primary subscription tier");
-                                        classification_error =
-                                            Some(err.to_string());
-                                    }
-                                }
-                            }
-                            result
-                        };
-
-                    if let Some((_, waiters)) = resolved_fetch {
-                        for sender in waiters {
-                            let response = match classification_error.as_ref() {
-                                Some(err) => Err(RemoteAccountProviderError::AccountResolutionsFailed(err.clone())),
-                                None => Ok(remote_account.clone()),
-                            };
-                            let _ = sender.send(response);
-                        }
-                    }
-
-                    if let Some(forward_update) = forward_update
-                        .filter(|_| classification_error.is_none())
-                    {
-                        if let Err(err) =
-                            subscription_forwarder.send(forward_update).await
-                        {
-                            warn!(
-                                pubkey = %update.pubkey,
-                                error = ?err,
-                                "Failed to forward subscription update"
-                            );
-                        }
-                    }
-                }
-            }
-        });
-        Ok(())
-    }
-
-    /// Convenience wrapper around [`RemoteAccountProvider::try_get_multi`] to fetch
-    /// a single account.
-    #[instrument(skip(self, fetch_context))]
-    pub async fn try_get(
-        &self,
-        pubkey: Pubkey,
-        fetch_context: impl Into<AccountFetchContext>,
-    ) -> RemoteAccountProviderResult<RemoteAccount> {
-        self.try_get_multi(&[pubkey], None, fetch_context, None)
-            .await
-            // SAFETY: we are guaranteed to have a single result here as
-            // otherwise we would have gotten an error
-            .map(|mut accs| accs.drain(..).next().unwrap())
-    }
-
-    #[instrument(skip(self, pubkeys, config, fetch_context))]
-    pub async fn try_get_multi_until_slots_match(
-        &self,
-        pubkeys: &[Pubkey],
-        config: Option<MatchSlotsConfig>,
-        fetch_context: impl Into<AccountFetchContext>,
-    ) -> RemoteAccountProviderResult<Vec<RemoteAccount>> {
-        use SlotsMatchResult::*;
-        let fetch_context = fetch_context.into();
-        let companion_fetch_kind =
-            config.as_ref().map(|config| config.companion_fetch_kind);
-        let config = config
-            .as_ref()
-            .map(MatchSlotsRetryConfig::from)
-            .unwrap_or_default();
-        let companion_fetch_started_at = std::time::Instant::now();
-        let mut companion_fetch_attempts = 1u64;
-        // 1. Fetch the _normal_ way and hope the slots match and if required
-        //    the min_context_slot is met
-        let mut remote_accounts = match self
-            .try_get_multi(pubkeys, None, fetch_context.clone(), None)
-            .await
-        {
-            Ok(accounts) => accounts,
-            Err(err) => {
-                observe_companion_fetch_if_configured(
-                    fetch_context.clone(),
-                    companion_fetch_kind,
-                    ChainlinkCompanionFetchOutcome::FailedRpc,
-                    companion_fetch_attempts,
-                    companion_fetch_started_at,
-                );
-                return Err(err);
-            }
-        };
-        // State observed at slot S must never be superseded by an older
-        // view: raise the floor to any found result already consumed.
-        let mut min_context_slot =
-            raised_min_context_slot(config.min_context_slot, &remote_accounts);
-        if let Match =
-            slots_match_and_meet_min_context(&remote_accounts, min_context_slot)
-        {
-            observe_companion_fetch_if_configured(
-                fetch_context.clone(),
-                companion_fetch_kind,
-                ChainlinkCompanionFetchOutcome::Succeeded,
-                companion_fetch_attempts,
-                companion_fetch_started_at,
-            );
-            return Ok(remote_accounts);
-        }
-
-        // Subscription results consumed to resolve this fetch must re-enter
-        // the update pipeline if the fetch fails, or they are lost.
-        let consumed_subscription_results: Vec<(Pubkey, RemoteAccount)> =
-            pubkeys
-                .iter()
-                .zip(&remote_accounts)
-                .filter(|(_, account)| {
-                    account.is_found()
-                        && account.source()
-                            == Some(RemoteAccountUpdateSource::Subscription)
-                })
-                .map(|(pubkey, account)| (*pubkey, account.clone()))
-                .collect();
-
-        // The fetch start slot honors the strictest floor observed so far:
-        // the caller's min_context_slot or any found result consumed above.
-        let mut fetch_start_slot = self
-            .chain_slot
-            .load()
-            .max(min_context_slot.unwrap_or_default());
-        // 2. Wait for the slots to match. Once the fast path mixed slots,
-        // retry with an RPC-only batch so all accounts share one response slot.
-        let start = std::time::Instant::now();
-        let mut retries = 0;
-        loop {
-            if tracing::enabled!(tracing::Level::TRACE) {
-                let slots = account_slots(&remote_accounts);
-                let pubkey_slots = pubkeys
-                    .iter()
-                    .zip(slots)
-                    .map(|(pk, slot)| format!("{pk}:{slot}"))
-                    .collect::<Vec<_>>()
-                    .join(", ");
-                trace!(
-                    "Retry({}) account fetch to sync non-matching slots [{}]",
-                    retries + 1,
-                    pubkey_slots
-                );
-            }
-            companion_fetch_attempts += 1;
-            remote_accounts = match self
-                .fetch_multi_rpc_only(
-                    pubkeys,
-                    fetch_start_slot,
-                    fetch_context.clone(),
-                )
-                .await
-            {
-                Ok(remote_accounts) => remote_accounts,
-                Err(err) => {
-                    let retry = next_match_slots_rpc_error_retry(
-                        &mut retries,
-                        start,
-                        &config,
-                    );
-                    debug!(
-                        pubkeys = %pubkeys_str(pubkeys),
-                        min_context_slot = ?min_context_slot,
-                        retries = retries,
-                        elapsed_ms = start.elapsed().as_millis() as u64,
-                        error = %err,
-                        "RPC-only account fetch failed while resolving accounts to compatible slots"
-                    );
-                    match retry {
-                        Ok(retry_delay) => {
-                            tokio::time::sleep(retry_delay).await;
-                            continue;
-                        }
-                        Err(_) => {
-                            observe_companion_fetch_if_configured(
-                                fetch_context,
-                                companion_fetch_kind,
-                                ChainlinkCompanionFetchOutcome::FailedRpc,
-                                companion_fetch_attempts,
-                                companion_fetch_started_at,
-                            );
-                            self.reforward_consumed_subscription_results(
-                                &consumed_subscription_results,
-                            );
-                            return Err(err);
-                        }
-                    }
-                }
-            };
-            for (pubkey, remote_account) in pubkeys.iter().zip(&remote_accounts)
-            {
-                let _subscription_guard =
-                    subscription_key_owned_guard_from_map(
-                        &self.subscription_key_locks,
-                        *pubkey,
-                    )
-                    .await;
-                if let Err(err) = self
-                    .subscription_tier_ctx()
-                    .apply_fetch_classification(
-                        pubkey,
-                        remote_account.slot(),
-                        !remote_account.is_found(),
-                    )
-                    .await
-                {
-                    self.reforward_consumed_subscription_results(
-                        &consumed_subscription_results,
-                    );
-                    return Err(err);
-                }
-            }
-            min_context_slot =
-                raised_min_context_slot(min_context_slot, &remote_accounts);
-            fetch_start_slot =
-                fetch_start_slot.max(min_context_slot.unwrap_or_default());
-            let slots_match_result = slots_match_and_meet_min_context(
-                &remote_accounts,
-                min_context_slot,
-            );
-            if let Match = slots_match_result {
-                observe_companion_fetch_if_configured(
-                    fetch_context.clone(),
-                    companion_fetch_kind,
-                    ChainlinkCompanionFetchOutcome::Succeeded,
-                    companion_fetch_attempts,
-                    companion_fetch_started_at,
-                );
-                return Ok(remote_accounts);
-            }
-
-            match next_match_slots_retry(&mut retries, start, &config) {
-                Ok(retry_delay) => {
-                    // If the slots don't match then wait for a bit and retry
-                    tokio::time::sleep(retry_delay).await;
-                    continue;
-                }
-                Err(limit) => {
-                    let remote_account_slots = account_slots(&remote_accounts);
-                    let remote_account_sources = remote_accounts
-                        .iter()
-                        .map(|account| account.source())
-                        .collect::<Vec<_>>();
-                    warn!(
-                        pubkeys = %pubkeys_str(pubkeys),
-                        slots = ?remote_account_slots,
-                        sources = ?remote_account_sources,
-                        min_context_slot = ?min_context_slot,
-                        retries = retries,
-                        elapsed_ms = start.elapsed().as_millis() as u64,
-                        limit = %limit,
-                        "Failed to resolve accounts to compatible slots"
-                    );
-                    match slots_match_result {
-                        // SAFETY: Match case is already handled and returns
-                        Match => unreachable!("we would have returned above"),
-                        Mismatch => {
-                            observe_companion_fetch_if_configured(
-                                fetch_context,
-                                companion_fetch_kind,
-                                ChainlinkCompanionFetchOutcome::FailedSlotMismatch,
-                                companion_fetch_attempts,
-                                companion_fetch_started_at,
-                            );
-                            self.reforward_consumed_subscription_results(
-                                &consumed_subscription_results,
-                            );
-                            return Err(
-                                RemoteAccountProviderError::SlotsDidNotMatch(
-                                    pubkeys_str(pubkeys),
-                                    remote_account_slots,
-                                    limit,
-                                ),
-                            );
-                        }
-                        MatchButBelowMinContextSlot(slot) => {
-                            observe_companion_fetch_if_configured(
-                                fetch_context,
-                                companion_fetch_kind,
-                                ChainlinkCompanionFetchOutcome::FailedMinContextSlot,
-                                companion_fetch_attempts,
-                                companion_fetch_started_at,
-                            );
-                            self.reforward_consumed_subscription_results(
-                                &consumed_subscription_results,
-                            );
-                            return Err(RemoteAccountProviderError::MatchingSlotsNotSatisfyingMinContextSlot(
-                                pubkeys_str(pubkeys),
-                                remote_account_slots,
-                                slot,
-                                limit,
-                            ));
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    /// Re-forwards found results consumed from the update pipeline by a
-    /// fetch that is now failing; otherwise the consumed update is lost.
-    /// Coalesced per account (newest slot wins) into an outbox drained by
-    /// [Self::start_replay_outbox_worker], so callers never block and the
-    /// backlog is bounded by the number of affected accounts.
-    fn reforward_consumed_subscription_results(
-        &self,
-        consumed: &[(Pubkey, RemoteAccount)],
-    ) {
-        if consumed.is_empty() {
-            return;
-        }
-        {
-            let mut outbox = self
-                .replay_outbox
-                .lock()
-                .unwrap_or_else(|poison| poison.into_inner());
-            for (pubkey, account) in consumed {
-                let entry = outbox.entry(*pubkey);
-                match entry {
-                    Entry::Occupied(mut existing)
-                        if existing.get().account.slot() < account.slot() =>
-                    {
-                        existing.insert(ForwardedSubscriptionUpdate {
-                            pubkey: *pubkey,
-                            account: account.clone(),
-                            source: SubscriptionSource::Replay,
-                        });
-                    }
-                    Entry::Occupied(_) => {}
-                    Entry::Vacant(vacant) => {
-                        vacant.insert(ForwardedSubscriptionUpdate {
-                            pubkey: *pubkey,
-                            account: account.clone(),
-                            source: SubscriptionSource::Replay,
-                        });
-                    }
-                }
-            }
-        }
-        self.replay_notify.notify_one();
-    }
-
-    /// Drains the replay outbox into the update pipeline. Runs detached so
-    /// replays never block a failing resolution; exits when the pipeline
-    /// closes.
-    fn start_replay_outbox_worker(&self) {
-        let outbox = Arc::clone(&self.replay_outbox);
-        let notify = Arc::clone(&self.replay_notify);
-        let forwarder = Arc::clone(&self.subscription_forwarder);
-        task::spawn(async move {
-            loop {
-                notify.notified().await;
-                loop {
-                    let update = {
-                        let mut outbox = outbox
-                            .lock()
-                            .unwrap_or_else(|poison| poison.into_inner());
-                        let Some(pubkey) = outbox.keys().next().copied() else {
-                            break;
-                        };
-                        outbox.remove(&pubkey)
-                    };
-                    let Some(update) = update else {
-                        break;
-                    };
-                    if forwarder.send(update).await.is_err() {
-                        return;
-                    }
-                }
-            }
-        });
-    }
-
-    /// Gets the accounts for the given pubkeys by fetching from RPC.
-    /// Always fetches fresh data. FetchCloner handles request deduplication.
-    /// Subscribes first to catch any updates that arrive during fetch.
-    #[instrument(skip(self, pubkeys, mark_empty_if_not_found, fetch_context))]
-    pub async fn try_get_multi(
-        &self,
-        pubkeys: &[Pubkey],
-        mark_empty_if_not_found: Option<&[Pubkey]>,
-        fetch_context: impl Into<AccountFetchContext>,
-        fetch_start_slot: Option<u64>,
-    ) -> RemoteAccountProviderResult<Vec<RemoteAccount>> {
-        if pubkeys.is_empty() {
-            return Ok(vec![]);
-        }
-        let fetch_context = fetch_context.into();
-
-        if tracing::enabled!(tracing::Level::TRACE) {
-            trace!("Fetching accounts");
-        }
-
-        let fetch_start_slot =
-            fetch_start_slot.unwrap_or_else(|| self.chain_slot.load());
-
-        // Receivers awaited by this call. One entry per input pubkey, in
-        // input order. Each receiver corresponds to a sender that was
-        // pushed into the FetchingAccountState.waiters queue (either
-        // because this call inserted the entry, or because it joined an
-        // existing in-flight fetch as a waiter).
-        let mut await_receivers: Vec<(
-            Pubkey,
-            oneshot::Receiver<FetchResult>,
-            PendingFetchWaiterGaugeGuard,
-        )> = Vec::with_capacity(pubkeys.len());
-
-        // Pubkeys this call actually inserted.
-        // Only these pubkeys cause side effects (subscription setup, fetch)
-        // by this call. Waiter-only pubkeys already have a subscription owned
-        // by the original claimer and are being fetched by them.
-        let mut claimed_pubkeys: Vec<Pubkey> = Vec::new();
-        let mut claimed_generations: HashMap<
-            Pubkey,
-            FetchingAccountGeneration,
-        > = HashMap::new();
-
-        {
-            let mut fetching = self
-                .fetching_accounts
-                .lock()
-                .unwrap_or_else(|poison| poison.into_inner());
-            for &pubkey in pubkeys {
-                let (sender, receiver) = oneshot::channel();
-                let mut claimed = false;
-                let layer = ChainlinkPendingFetchLayer::RemoteAccountProvider;
-                let mut waiter_guard =
-                    PendingFetchWaiterGaugeGuard::inactive(layer);
-                match fetching.entry(pubkey) {
-                    Entry::Occupied(mut entry) => {
-                        entry.get_mut().waiters.push(sender);
-                        inc_chainlink_pending_fetch_accounts_with_context(
-                            fetch_context.clone(),
-                            layer,
-                            ChainlinkPendingFetchOutcome::JoinedExisting,
-                            1,
-                        );
-                        inc_chainlink_pending_fetch_waiters_with_context(
-                            fetch_context.clone(),
-                            layer,
-                            1,
-                        );
-                        inc_chainlink_pending_fetch_waiters_gauge(layer);
-                        waiter_guard =
-                            PendingFetchWaiterGaugeGuard::active(layer);
-                    }
-                    Entry::Vacant(entry) => {
-                        let generation =
-                            self.next_fetching_account_generation();
-                        entry.insert(FetchingAccountState {
-                            generation,
-                            fetch_start_slot,
-                            fetch_context: fetch_context.clone(),
-                            owner_started_at: std::time::Instant::now(),
-                            waiters: vec![sender],
-                        });
-                        inc_chainlink_pending_fetch_accounts_with_context(
-                            fetch_context.clone(),
-                            layer,
-                            ChainlinkPendingFetchOutcome::Owned,
-                            1,
-                        );
-                        claimed_generations.insert(pubkey, generation);
-                        claimed = true;
-                    }
-                }
-                if claimed {
-                    claimed_pubkeys.push(pubkey);
-                }
-                await_receivers.push((pubkey, receiver, waiter_guard));
-            }
-        }
-
-        if fetch_context.should_count_remote_account_claims()
-            && !claimed_pubkeys.is_empty()
-        {
-            let unique_claimed_pubkey_count = claimed_pubkeys
-                .iter()
-                .copied()
-                .collect::<HashSet<_>>()
-                .len();
-            fetch_context
-                .add_remote_account_claims(unique_claimed_pubkey_count);
-        }
-
-        // Setup subscriptions and trigger the fetch only for pubkeys this
-        // call actually claimed. Waiter-only pubkeys already have a
-        // subscription and an in-flight fetch owned by the original
-        // claimer; doing it again here would duplicate work and (for
-        // setup_subscriptions) double-count subscription side effects.
-        if !claimed_pubkeys.is_empty() {
-            let mut subscription_setup_guard =
-                ClaimedSubscriptionSetupGuard::new(
-                    self.fetching_accounts.clone(),
-                    self.subscription_ownership.clone(),
-                    self.subscription_transition_lock.clone(),
-                    self.lrucache_subscribed_accounts.clone(),
-                    self.secondary_subscriptions.clone(),
-                    claimed_pubkeys.clone(),
-                    claimed_generations.clone(),
-                );
-            if let Err(err) = self
-                .setup_subscriptions(&claimed_pubkeys, fetch_context.clone())
-                .await
-            {
-                subscription_setup_guard
-                    .cleanup_with_error(err.to_string())
-                    .await;
-                return Err(err);
-            }
-            subscription_setup_guard.disarm();
-
-            // Start the fetch for the claimed pubkeys only. Claim sets
-            // above the RPC limit are split into evenly sized chunks
-            let min_context_slot = fetch_start_slot;
-            for chunk in balanced_chunks(claimed_pubkeys) {
-                self.fetch(
-                    chunk,
-                    claimed_generations.clone(),
-                    mark_empty_if_not_found,
-                    min_context_slot,
-                    fetch_context.clone(),
-                );
-            }
-        }
-
-        // Wait for all accounts to resolve (either from fetch or
-        // subscription override). We await receivers in input pubkey
-        // order so the returned Vec is index-aligned with `pubkeys`.
-        let mut resolved_accounts = vec![];
-        let mut errors = vec![];
-
-        for (idx, (pubkey, receiver, mut waiter_guard)) in
-            await_receivers.into_iter().enumerate()
-        {
-            let receiver_result = receiver.await;
-            waiter_guard.finish();
-            match receiver_result {
-                Ok(result) => match result {
-                    Ok(remote_account) => {
-                        resolved_accounts.push(remote_account)
-                    }
-                    Err(err) => {
-                        warn!(pubkey = %pubkey, error = %err, "Failed to fetch account");
-                        errors.push((idx, err));
-                    }
-                },
-                Err(err) => {
-                    warn!(pubkey = %pubkey, stream_index = idx, error = ?err, total_pubkeys = pubkeys.len(), "Failed to resolve account (unexpected RecvError)");
-                    errors.push((
-                        idx,
-                        RemoteAccountProviderError::RecvrError(err),
-                    ));
-                }
-            }
-        }
-
-        if errors.is_empty() {
-            assert_eq!(
-                resolved_accounts.len(),
-                pubkeys.len(),
-                "BUG: resolved accounts and pubkeys length mismatch"
-            );
-            Ok(resolved_accounts)
-        } else {
-            Err(RemoteAccountProviderError::AccountResolutionsFailed(
-                errors
-                    .iter()
-                    .map(|(idx, err)| {
-                        let pubkey = pubkeys
-                            .get(*idx)
-                            .map(|pk| pk.to_string())
-                            .unwrap_or_else(|| {
-                                "BUG: could not match pubkey".to_string()
-                            });
-                        format!("{pubkey}: {err:?}")
-                    })
-                    .collect::<Vec<_>>()
-                    .join(",\n"),
-            ))
-        }
-    }
-
-    async fn fetch_multi_rpc_only(
-        &self,
-        pubkeys: &[Pubkey],
-        min_context_slot: u64,
-        fetch_context: AccountFetchContext,
-    ) -> RemoteAccountProviderResult<Vec<RemoteAccount>> {
-        // This must stay a single wire call so all results share one
-        // response slot (the slot-match contract callers verify);
-        // slot-consistent sets must fit within the RPC limit.
-        debug_assert!(
-            pubkeys.len() <= MAX_MULTIPLE_ACCOUNTS_PER_REQUEST,
-            "fetch_multi_rpc_only cannot chunk {} keys (limit {})",
-            pubkeys.len(),
-            MAX_MULTIPLE_ACCOUNTS_PER_REQUEST
-        );
-        let config = RpcAccountInfoConfig {
-            commitment: Some(self.rpc_client.commitment()),
-            min_context_slot: Some(min_context_slot),
-            encoding: Some(UiAccountEncoding::Base64Zstd),
-            data_slice: None,
-        };
-
-        metrics::inc_remote_account_provider_a_count();
-        let response = tokio::time::timeout(RPC_FETCH_TIMEOUT, async {
-            self.rpc_client
-                .get_multiple_accounts_with_config(pubkeys, config)
-                .await
-        })
-        .await
-        .map_err(|_| {
-            RemoteAccountProviderError::AccountResolutionsFailed(format!(
-                "RPC call timeout fetching accounts {} after {}ms",
-                pubkeys_str(pubkeys),
-                RPC_FETCH_TIMEOUT.as_millis()
-            ))
-        })?
-        .map_err(|err| {
-            RemoteAccountProviderError::AccountResolutionsFailed(format!(
-                "RpcError fetching accounts {}: {err:?}",
-                pubkeys_str(pubkeys)
-            ))
-        })?;
-
-        let response_slot = response.context.slot;
-        if response_slot < min_context_slot {
-            return Err(RemoteAccountProviderError::AccountResolutionsFailed(
-                format!(
-                    "Response slot {response_slot} < {min_context_slot} fetching accounts {}",
-                    pubkeys_str(pubkeys)
-                ),
-            ));
-        }
-
-        let response_value_len = response.value.len();
-        if response_value_len != pubkeys.len() {
-            return Err(RemoteAccountProviderError::AccountResolutionsFailed(
-                format!(
-                    "RPC returned {response_value_len} account results for {} requested accounts: {}",
-                    pubkeys.len(),
-                    pubkeys_str(pubkeys)
-                ),
-            ));
-        }
-
-        let mut found_count = 0u64;
-        let mut not_found_count = 0u64;
-        let remote_accounts = response
-            .value
-            .into_iter()
-            .map(|account| match account {
-                Some(account) => {
-                    found_count += 1;
-                    RemoteAccount::from_fresh_account(
-                        account,
-                        response_slot,
-                        RemoteAccountUpdateSource::Fetch,
-                    )
-                }
-                None => {
-                    not_found_count += 1;
-                    RemoteAccount::NotFound(response_slot)
-                }
-            })
-            .collect();
-
-        inc_account_fetches_success(pubkeys.len() as u64);
-        inc_account_fetches_found_with_context(
-            fetch_context.clone(),
-            found_count,
-        );
-        inc_account_fetches_not_found_with_context(
-            fetch_context,
-            not_found_count,
-        );
-
-        Ok(remote_accounts)
-    }
-
-    async fn setup_subscriptions(
-        &self,
-        pubkeys: &[Pubkey],
-        fetch_context: AccountFetchContext,
-    ) -> RemoteAccountProviderResult<()> {
-        if tracing::enabled!(tracing::Level::TRACE) {
-            let pubkeys_str = pubkeys
-                .iter()
-                .map(|pk| pk.to_string())
-                .collect::<Vec<_>>()
-                .join(", ");
-            trace!(pubkeys = pubkeys_str, "Subscribing to accounts");
-        }
-        // Send all subscription requests in parallel (non-fail-fast).
-        // We use join_all instead of try_join_all to ensure ALL acquire
-        // attempts complete, even if some fail.
-        let subscription_results = join_all(pubkeys.iter().map(|pubkey| {
-            let fetch_context = fetch_context.clone();
-            async move {
-                self.acquire_subscription_with_origin(
-                    pubkey,
-                    SubscriptionReason::DirectAccount,
-                    SubscriptionRegistrationOrigin::Fetch(fetch_context),
-                )
-                .await
-            }
-        }))
-        .await;
-
-        let mut errors = Vec::new();
-        let mut acquired = Vec::new();
-        for (result, pubkey) in
-            subscription_results.into_iter().zip(pubkeys.iter())
-        {
-            match result {
-                Err(err) => {
-                    error!(
-                        pubkey = %pubkey, err = ?err,
-                        "Failed to subscribe to account"
-                    );
-                    errors.push((*pubkey, err));
-                }
-                Ok(()) => acquired.push(*pubkey),
-            }
-        }
-
-        if !errors.is_empty() {
-            for pubkey in &acquired {
-                if let Err(unsub_err) = self
-                    .release_single_subscription(
-                        pubkey,
-                        SubscriptionReason::DirectAccount,
-                    )
-                    .await
-                {
-                    if matches!(
-                        unsub_err,
-                        RemoteAccountProviderError::AccountSubscriptionDoesNotExist(_)
-                    ) {
-                        debug!(
-                            pubkey = %pubkey, err = ?unsub_err,
-                            "Failed to unsubscribe after partial \
-                             subscription failure"
-                        );
-                    } else {
-                        warn!(
-                            pubkey = %pubkey, err = ?unsub_err,
-                            "Failed to unsubscribe after partial \
-                             subscription failure"
-                        );
-                    }
-                }
-            }
-            // A single failure keeps its type so callers can react to
-            // specific variants (e.g. capacity exhaustion).
-            if errors.len() == 1 {
-                // SAFETY: len checked above
-                return Err(errors.pop().unwrap().1);
-            }
-            return Err(
-                RemoteAccountProviderError::AccountSubscriptionsTaskFailed(
-                    format!(
-                        "{} subscription(s) failed: [{}]",
-                        errors.len(),
-                        errors
-                            .iter()
-                            .map(|(pubkey, err)| format!("{pubkey}: {err}"))
-                            .collect::<Vec<_>>()
-                            .join(", ")
-                    ),
-                ),
-            );
-        }
-
-        Ok(())
-    }
-
-    /// Registers a new subscription for the given pubkey.
-    async fn register_subscription(
-        &self,
-        pubkey: &Pubkey,
-        reason: SubscriptionReason,
-        origin: SubscriptionRegistrationOrigin,
-    ) -> RemoteAccountProviderResult<()> {
-        if matches!(origin, SubscriptionRegistrationOrigin::Fetch(_))
-            && reason == SubscriptionReason::DirectAccount
-            && self.lrucache_subscribed_accounts.can_evict(pubkey)
-        {
-            return self
-                .subscription_tier_ctx()
-                .register_secondary(pubkey, reason, origin)
-                .await;
-        }
-
-        let tier_ctx = self.subscription_tier_ctx();
-        let has_capacity = {
-            let _transition_guard =
-                self.subscription_transition_lock.lock().await;
-            tier_ctx
-                .has_capacity_with_protection(
-                    &self.lrucache_subscribed_accounts,
-                    pubkey,
-                )
-                .await
-        };
-        if !has_capacity {
-            inc_chainlink_subscription_registration_accounts(
-                origin,
-                reason.into(),
-                SubscriptionRegistrationOutcome::RejectedNoCapacity,
-            );
-            return Err(
-                RemoteAccountProviderError::NoEvictableSubscriptionCapacity {
-                    pubkey: *pubkey,
-                },
-            );
-        }
-
-        // 1. First realize subscription. Runs outside the transition lock;
-        // the per-key guard held by the caller serializes this key.
-        if let Err(err) = self.pubsub_client.subscribe(*pubkey, None).await {
-            inc_chainlink_subscription_registration_accounts(
-                origin,
-                reason.into(),
-                SubscriptionRegistrationOutcome::SubscribeError,
-            );
-            return Err(err);
-        }
-
-        // 2. Add to LRU cache
-        // If an account is evicted then we need to unsubscribe from it
-        // and then inform upstream that we are no longer tracking it
-        let add_outcome = {
-            let _transition_guard =
-                self.subscription_transition_lock.lock().await;
-            let add_outcome = tier_ctx
-                .add_with_protection(
-                    &self.lrucache_subscribed_accounts,
-                    *pubkey,
-                )
-                .await;
-            if !matches!(add_outcome, AddAccountOutcome::NoEvictableCandidate) {
-                self.remove_from_secondary(pubkey);
-            }
-            add_outcome
-        };
-
-        match add_outcome {
-            AddAccountOutcome::AlreadyPresent => {
-                inc_chainlink_subscription_registration_accounts(
-                    origin,
-                    reason.into(),
-                    SubscriptionRegistrationOutcome::AlreadyPresent,
-                );
-            }
-            AddAccountOutcome::Added => {
-                inc_chainlink_subscription_registration_accounts(
-                    origin,
-                    reason.into(),
-                    SubscriptionRegistrationOutcome::AddedBelowCapacity,
-                );
-            }
-            AddAccountOutcome::Evicted(evicted) => {
-                trace!(evicted = %evicted, "Evicting account");
-                tier_ctx.spawn_evicted_cleanup(evicted);
-                inc_chainlink_subscription_registration_accounts(
-                    origin,
-                    reason.into(),
-                    SubscriptionRegistrationOutcome::EvictedCandidate,
-                );
-            }
-            AddAccountOutcome::NoEvictableCandidate => {
-                tier_ctx.cleanup_rejected_subscription(*pubkey).await?;
-                debug!(
-                    pubkey = %pubkey,
-                    "No evictable subscription capacity available; all LRU candidates are protected"
-                );
-                inc_chainlink_subscription_registration_accounts(
-                    origin,
-                    reason.into(),
-                    SubscriptionRegistrationOutcome::RejectedAndUnsubscribed,
-                );
-                return Err(
-                    RemoteAccountProviderError::NoEvictableSubscriptionCapacity {
-                        pubkey: *pubkey,
-                    },
-                );
-            }
-        }
-
-        Ok(())
-    }
-
-    pub(crate) async fn send_removal_update(
-        &self,
-        evicted: Pubkey,
-    ) -> RemoteAccountProviderResult<()> {
-        self.removed_account_tx.send(evicted).await.map_err(
-            RemoteAccountProviderError::FailedToSendAccountRemovalUpdate,
-        )?;
-        Ok(())
-    }
-
-    /// Check if an account is currently being watched (subscribed to)
-    /// This does not consider accounts like the clock sysvar that are watched as
-    /// part of the provider's internal logic.
-    pub fn is_watching(&self, pubkey: &Pubkey) -> bool {
-        self.lrucache_subscribed_accounts.contains(pubkey)
-            || self.secondary_subscriptions.contains(pubkey)
-    }
-
-    /// Removes a pubkey from the secondary LRU; safe for never-evict keys.
-    fn remove_from_secondary(&self, pubkey: &Pubkey) {
-        if self.secondary_subscriptions.contains(pubkey) {
-            self.secondary_subscriptions.remove(pubkey);
-        }
-        self.confirmed_missing_subscriptions
-            .lock()
-            .unwrap_or_else(|poison| poison.into_inner())
-            .remove(pubkey);
-    }
-
-    fn subscription_tier_ctx(&self) -> SubscriptionTierCtx<U> {
-        SubscriptionTierCtx {
-            primary: self.lrucache_subscribed_accounts.clone(),
-            secondary: self.secondary_subscriptions.clone(),
-            pubsub_client: self.pubsub_client.clone(),
-            subscription_ownership: self.subscription_ownership.clone(),
-            subscription_transition_lock: self
-                .subscription_transition_lock
-                .clone(),
-            subscription_key_locks: self.subscription_key_locks.clone(),
-            fetching_accounts: self.fetching_accounts.clone(),
-            capacity_eviction_protection: self
-                .capacity_eviction_protection
-                .clone(),
-            confirmed_missing_subscriptions: self
-                .confirmed_missing_subscriptions
-                .clone(),
-            removed_account_tx: self.removed_account_tx.clone(),
-        }
-    }
-
-    pub(crate) async fn evict_unwatched_with_subscription_lock<F, Fut>(
-        &self,
-        pubkey: &Pubkey,
-        evict: F,
-    ) -> bool
-    where
-        F: FnOnce() -> Fut,
-        Fut: std::future::Future<Output = ()>,
-    {
-        let subscription_key_lock = self.subscription_key_lock(pubkey).await;
-        let _subscription_guard = subscription_key_lock.lock().await;
-
-        if self.is_watching(pubkey) {
-            return false;
-        }
-
-        evict().await;
-        true
-    }
-
-    async fn subscription_key_lock(
-        &self,
-        pubkey: &Pubkey,
-    ) -> Arc<AsyncMutex<()>> {
-        subscription_key_lock_from_map(&self.subscription_key_locks, pubkey)
-            .await
-    }
-
-    pub async fn acquire_subscription(
-        &self,
-        pubkey: &Pubkey,
-        reason: SubscriptionReason,
-    ) -> RemoteAccountProviderResult<()> {
-        self.acquire_subscription_with_mode(
-            pubkey,
-            reason,
-            false,
-            SubscriptionRegistrationOrigin::Internal,
-        )
-        .await
-    }
-
-    async fn acquire_subscription_with_origin(
-        &self,
-        pubkey: &Pubkey,
-        reason: SubscriptionReason,
-        origin: SubscriptionRegistrationOrigin,
-    ) -> RemoteAccountProviderResult<()> {
-        self.acquire_subscription_with_mode(pubkey, reason, false, origin)
-            .await
-    }
-
-    pub async fn ensure_subscription(
-        &self,
-        pubkey: &Pubkey,
-        reason: SubscriptionReason,
-    ) -> RemoteAccountProviderResult<()> {
-        self.acquire_subscription_with_mode(
-            pubkey,
-            reason,
-            true,
-            SubscriptionRegistrationOrigin::Internal,
-        )
-        .await
-    }
-
-    async fn acquire_subscription_with_mode(
-        &self,
-        pubkey: &Pubkey,
-        reason: SubscriptionReason,
-        skip_existing_reason: bool,
-        origin: SubscriptionRegistrationOrigin,
-    ) -> RemoteAccountProviderResult<()> {
-        // The per-key guard serializes every transition of this key,
-        // including the network calls made below. The transition lock is
-        // acquired only inside the tier-state helpers.
-        let subscription_key_lock = self.subscription_key_lock(pubkey).await;
-        let _subscription_guard = subscription_key_lock.lock().await;
-
-        let mut ownership = self.subscription_ownership.lock().await;
-        if let Some(existing) = ownership.get_mut(pubkey) {
-            let classification_placeholder_generation =
-                existing.classification_placeholder_generation;
-            let acquired_reason =
-                !skip_existing_reason || !existing.contains(reason);
-            if acquired_reason {
-                existing.acquire(reason);
-            }
-            drop(ownership);
-
-            let repair_result = if self
-                .lrucache_subscribed_accounts
-                .contains(pubkey)
-            {
-                self.lrucache_subscribed_accounts.promote_multi(&[pubkey]);
-                Ok(())
-            } else if self.secondary_subscriptions.contains(pubkey) {
-                self.secondary_subscriptions.promote_multi(&[pubkey]);
-                let keep_secondary =
-                    matches!(origin, SubscriptionRegistrationOrigin::Fetch(_))
-                        && reason == SubscriptionReason::DirectAccount;
-                if !keep_secondary {
-                    match self
-                        .subscription_tier_ctx()
-                        .try_promote_found_to_primary(*pubkey, true)
-                        .await
-                    {
-                        Ok(PromotionOutcome::Promoted) => Ok(()),
-                        // Promoted by another transition mid-flight; the key
-                        // holds primary membership and the reason stands.
-                        Ok(PromotionOutcome::NotInSecondary) => Ok(()),
-                        // Evicted by another key's admission mid-flight;
-                        // register it from scratch.
-                        Ok(PromotionOutcome::Evicted) => {
-                            self.register_subscription(pubkey, reason, origin.clone())
-                                .await
-                        }
-                        Ok(PromotionOutcome::NoCapacity)
-                            if reason
-                                == SubscriptionReason::UndelegationTracking =>
-                        {
-                            Ok(())
-                        }
-                        Ok(PromotionOutcome::NoCapacity) => Err(
-                            RemoteAccountProviderError::NoEvictableSubscriptionCapacity {
-                                pubkey: *pubkey,
-                            },
-                        ),
-                        Err(err) => Err(err),
-                    }
-                } else {
-                    let confirmed_missing = self
-                        .confirmed_missing_subscriptions
-                        .lock()
-                        .unwrap_or_else(|poison| poison.into_inner())
-                        .contains(pubkey);
-                    if confirmed_missing {
-                        // Flow the error into repair_result (no early return)
-                        // so the acquired-reason rollback below executes.
-                        match self.pubsub_client.subscribe(*pubkey, None).await
-                        {
-                            Ok(()) => {
-                                self.confirmed_missing_subscriptions
-                                    .lock()
-                                    .unwrap_or_else(|poison| {
-                                        poison.into_inner()
-                                    })
-                                    .remove(pubkey);
-                                Ok(())
-                            }
-                            Err(err) => Err(err),
-                        }
-                    } else {
-                        Ok(())
-                    }
-                }
-            } else {
-                self.register_subscription(pubkey, reason, origin.clone())
-                    .await
-            };
-
-            if let Err(err) = repair_result {
-                if acquired_reason {
-                    if let Some(existing) =
-                        self.subscription_ownership.lock().await.get_mut(pubkey)
-                    {
-                        if existing.release(reason) {
-                            existing.classification_placeholder_generation =
-                                classification_placeholder_generation;
-                        }
-                    }
-                }
-                return Err(err);
-            }
-            inc_chainlink_subscription_registration_accounts(
-                origin.clone(),
-                reason.into(),
-                SubscriptionRegistrationOutcome::AlreadyPresent,
-            );
-            return Ok(());
-        }
-        drop(ownership);
-
-        self.register_subscription(pubkey, reason, origin.clone())
-            .await?;
-
-        let mut ownership = self.subscription_ownership.lock().await;
-        ownership.entry(*pubkey).or_default().acquire(reason);
-        Ok(())
-    }
-
-    pub async fn release_single_subscription(
-        &self,
-        pubkey: &Pubkey,
-        reason: SubscriptionReason,
-    ) -> RemoteAccountProviderResult<bool> {
-        self.release_subscription_with_mode(
-            pubkey,
-            reason,
-            SubscriptionReleaseMode::Single,
-        )
-        .await
-    }
-
-    pub(crate) async fn release_subscription_with_mode(
-        &self,
-        pubkey: &Pubkey,
-        reason: SubscriptionReason,
-        mode: SubscriptionReleaseMode,
-    ) -> RemoteAccountProviderResult<bool> {
-        let subscription_key_lock = self.subscription_key_lock(pubkey).await;
-        let _subscription_guard = subscription_key_lock.lock().await;
-
-        if !self.lrucache_subscribed_accounts.can_evict(pubkey) {
-            inc_chainlink_subscription_release_accounts(
-                reason.into(),
-                SubscriptionReleaseOutcome::RetainedIntentionally,
-            );
-            return Ok(false);
-        }
-
-        let released_count = {
-            let mut ownership = self.subscription_ownership.lock().await;
-            let (is_empty, released_count) = match ownership.get_mut(pubkey) {
-                Some(existing) => {
-                    let released_count = match mode {
-                        SubscriptionReleaseMode::Single => {
-                            existing.release(reason);
-                            1
-                        }
-                        SubscriptionReleaseMode::All => {
-                            existing.release_all(reason)
-                        }
-                    };
-                    (existing.is_empty(), released_count)
-                }
-                None => {
-                    inc_chainlink_subscription_release_accounts(
-                        reason.into(),
-                        SubscriptionReleaseOutcome::AlreadyAbsent,
-                    );
-                    return Ok(false);
-                }
-            };
-            if !is_empty {
-                inc_chainlink_subscription_release_accounts(
-                    reason.into(),
-                    SubscriptionReleaseOutcome::RetainedOtherReasons,
-                );
-                return Ok(false);
-            }
-            released_count
-        };
-
-        let success = subscription_reconciler::unsubscribe_and_notify_removal(
-            *pubkey,
-            &self.pubsub_client,
-            &self.removed_account_tx,
-            SubscriptionCleanupSource::NormalRelease,
-        )
-        .await;
-
-        if success {
-            inc_chainlink_subscription_release_accounts(
-                reason.into(),
-                SubscriptionReleaseOutcome::Unsubscribed,
-            );
-            let _transition_guard =
-                self.subscription_transition_lock.lock().await;
-            self.subscription_ownership.lock().await.remove(pubkey);
-            self.lrucache_subscribed_accounts.remove(pubkey);
-            self.remove_from_secondary(pubkey);
-        } else {
-            inc_chainlink_subscription_release_accounts(
-                reason.into(),
-                SubscriptionReleaseOutcome::UnsubscribeFailed,
-            );
-            let mut ownership = self.subscription_ownership.lock().await;
-            if ownership
-                .get(pubkey)
-                .is_none_or(SubscriptionOwnership::is_empty)
-            {
-                let ownership = ownership.entry(*pubkey).or_default();
-                for _ in 0..released_count {
-                    ownership.acquire(reason);
-                }
-            }
-        }
-
-        Ok(success)
-    }
-
-    pub(crate) async fn release_subscription_reason_silently_for_delegated_account(
-        &self,
-        pubkey: &Pubkey,
-        reason: SubscriptionReason,
-    ) -> RemoteAccountProviderResult<bool> {
-        let subscription_key_lock = self.subscription_key_lock(pubkey).await;
-        let _subscription_guard = subscription_key_lock.lock().await;
-
-        let released_count = {
-            let release_mode = SubscriptionReleaseMode::All;
-            let mut ownership = self.subscription_ownership.lock().await;
-            let (is_empty, released_count) = match ownership.get_mut(pubkey) {
-                Some(existing) => {
-                    let released_count = match release_mode {
-                        SubscriptionReleaseMode::Single => {
-                            existing.release(reason);
-                            1
-                        }
-                        SubscriptionReleaseMode::All => {
-                            existing.release_all(reason)
-                        }
-                    };
-                    (existing.is_empty(), released_count)
-                }
-                None => {
-                    inc_chainlink_subscription_release_accounts(
-                        reason.into(),
-                        SubscriptionReleaseOutcome::AlreadyAbsent,
-                    );
-                    inc_chainlink_subscription_cleanup_accounts(
-                        SubscriptionCleanupSource::DelegatedAccountSilent,
-                        SubscriptionCleanupOutcome::AlreadyAbsent,
-                    );
-                    return Ok(false);
-                }
-            };
-
-            if released_count == 0 {
-                inc_chainlink_subscription_release_accounts(
-                    reason.into(),
-                    SubscriptionReleaseOutcome::AlreadyAbsent,
-                );
-                inc_chainlink_subscription_cleanup_accounts(
-                    SubscriptionCleanupSource::DelegatedAccountSilent,
-                    SubscriptionCleanupOutcome::AlreadyAbsent,
-                );
-                return Ok(false);
-            }
-
-            if !is_empty {
-                inc_chainlink_subscription_release_accounts(
-                    reason.into(),
-                    SubscriptionReleaseOutcome::RetainedOtherReasons,
-                );
-                inc_chainlink_subscription_cleanup_accounts(
-                    SubscriptionCleanupSource::DelegatedAccountSilent,
-                    SubscriptionCleanupOutcome::RetainedIntentionally,
-                );
-                trace!(
-                    pubkey = %pubkey,
-                    ?reason,
-                    released_count,
-                    "Released delegated-account subscription ownership; \
-                     kept protected/live subscription and LRU entry"
-                );
-                return Ok(false);
-            }
-
-            released_count
-        };
-
-        match self.pubsub_client.unsubscribe(*pubkey).await {
-            Ok(()) => {
-                inc_chainlink_subscription_release_accounts(
-                    reason.into(),
-                    SubscriptionReleaseOutcome::Unsubscribed,
-                );
-                inc_chainlink_subscription_cleanup_accounts(
-                    SubscriptionCleanupSource::DelegatedAccountSilent,
-                    SubscriptionCleanupOutcome::Unsubscribed,
-                );
-                {
-                    let _transition_guard =
-                        self.subscription_transition_lock.lock().await;
-                    self.subscription_ownership.lock().await.remove(pubkey);
-                    self.lrucache_subscribed_accounts.remove(pubkey);
-                    self.remove_from_secondary(pubkey);
-                }
-                trace!(
-                    pubkey = %pubkey,
-                    ?reason,
-                    "Removed final delegated-account subscription ownership \
-                     and LRU entry silently; no removal notification emitted"
-                );
-                Ok(true)
-            }
-            Err(err) => {
-                if matches!(
-                    err,
-                    RemoteAccountProviderError::AccountSubscriptionDoesNotExist(
-                        _
-                    )
-                ) {
-                    inc_chainlink_subscription_release_accounts(
-                        reason.into(),
-                        SubscriptionReleaseOutcome::AlreadyAbsent,
-                    );
-                    inc_chainlink_subscription_cleanup_accounts(
-                        SubscriptionCleanupSource::DelegatedAccountSilent,
-                        SubscriptionCleanupOutcome::AlreadyAbsent,
-                    );
-                    {
-                        let _transition_guard =
-                            self.subscription_transition_lock.lock().await;
-                        self.subscription_ownership.lock().await.remove(pubkey);
-                        self.lrucache_subscribed_accounts.remove(pubkey);
-                        self.remove_from_secondary(pubkey);
-                    }
-                    trace!(
-                        pubkey = %pubkey,
-                        ?reason,
-                        "Removed stale delegated-account LRU entry for missing subscription"
-                    );
-                    return Ok(false);
-                }
-
-                inc_chainlink_subscription_release_accounts(
-                    reason.into(),
-                    SubscriptionReleaseOutcome::UnsubscribeFailed,
-                );
-                inc_chainlink_subscription_cleanup_accounts(
-                    SubscriptionCleanupSource::DelegatedAccountSilent,
-                    SubscriptionCleanupOutcome::UnsubscribeFailed,
-                );
-                let mut ownership = self.subscription_ownership.lock().await;
-                if ownership
-                    .get(pubkey)
-                    .is_none_or(SubscriptionOwnership::is_empty)
-                {
-                    let ownership = ownership.entry(*pubkey).or_default();
-                    for _ in 0..released_count {
-                        ownership.acquire(reason);
-                    }
-                }
-                drop(ownership);
-
-                Err(err)
-            }
-        }
-    }
-
-    /// Subscribe to program account updates
-    #[instrument(skip(self))]
-    pub async fn subscribe_program(
-        &self,
-        program_id: Pubkey,
-    ) -> RemoteAccountProviderResult<()> {
-        self.pubsub_client.subscribe_program(program_id).await
-    }
-
-    /// Unsubscribe from an account
-    #[instrument(skip(self))]
-    pub async fn unsubscribe(
-        &self,
-        pubkey: &Pubkey,
-    ) -> RemoteAccountProviderResult<()> {
-        let subscription_key_lock = self.subscription_key_lock(pubkey).await;
-        let _subscription_guard = subscription_key_lock.lock().await;
-
-        if !self.lrucache_subscribed_accounts.can_evict(pubkey) {
-            warn!(pubkey = %pubkey, "Tried to unsubscribe from account that should never be evicted");
-            inc_chainlink_subscription_cleanup_accounts(
-                SubscriptionCleanupSource::ManualUnsubscribe,
-                SubscriptionCleanupOutcome::RetainedIntentionally,
-            );
-            return Ok(());
-        }
-
-        if !self.lrucache_subscribed_accounts.contains(pubkey)
-            && !self.secondary_subscriptions.contains(pubkey)
-        {
-            trace!(pubkey = %pubkey, "Already unsubscribed from LRU");
-            inc_chainlink_subscription_cleanup_accounts(
-                SubscriptionCleanupSource::ManualUnsubscribe,
-                SubscriptionCleanupOutcome::AlreadyAbsent,
-            );
-            return Ok(());
-        }
-
-        let success = subscription_reconciler::unsubscribe_and_notify_removal(
-            *pubkey,
-            &self.pubsub_client,
-            &self.removed_account_tx,
-            SubscriptionCleanupSource::ManualUnsubscribe,
-        )
-        .await;
-
-        if success {
-            let _transition_guard =
-                self.subscription_transition_lock.lock().await;
-            self.lrucache_subscribed_accounts.remove(pubkey);
-            self.remove_from_secondary(pubkey);
-            self.subscription_ownership.lock().await.remove(pubkey);
-        }
-
-        Ok(())
-    }
-
-    /// Fetches a byte range of an account via `dataSlice`, retrying until the
-    /// response context reaches `min_context_slot` so the caller never acts on
-    /// pre-notification state. Returns `None` if the account does not exist.
-    pub(crate) async fn get_account_data_slice(
-        &self,
-        pubkey: &Pubkey,
-        offset: usize,
-        length: usize,
-        min_context_slot: u64,
-    ) -> RemoteAccountProviderResult<Option<Vec<u8>>> {
-        let mut last_err = String::new();
-        for attempt in 1..=DATA_SLICE_FETCH_MAX_ATTEMPTS {
-            let config = RpcAccountInfoConfig {
-                commitment: Some(self.rpc_client.commitment()),
-                // The context slot is verified locally below; passing
-                // min_context_slot would surface provider-specific errors
-                // instead of a plain lagging context.
-                min_context_slot: None,
-                encoding: Some(UiAccountEncoding::Base64),
-                data_slice: Some(UiDataSliceConfig { offset, length }),
-            };
-            match tokio::time::timeout(
-                RPC_FETCH_TIMEOUT,
-                self.rpc_client.get_account_with_config(pubkey, config),
-            )
-            .await
-            {
-                Ok(Ok(response)) => {
-                    if response.context.slot >= min_context_slot {
-                        return Ok(response.value.map(|account| account.data));
-                    }
-                    last_err = format!(
-                        "context slot {} below min {}",
-                        response.context.slot, min_context_slot
-                    );
-                }
-                Ok(Err(err)) => last_err = format!("{err:?}"),
-                Err(_) => {
-                    last_err = format!(
-                        "timeout after {}ms",
-                        RPC_FETCH_TIMEOUT.as_millis()
-                    )
-                }
-            }
-            if attempt < DATA_SLICE_FETCH_MAX_ATTEMPTS {
-                tokio::time::sleep(RPC_FETCH_RETRY_DELAY).await;
-            }
-        }
-        Err(RemoteAccountProviderError::AccountDataSliceFetchFailed(
-            *pubkey,
-            min_context_slot,
-            last_err,
-        ))
-    }
-
-    /// Tries to fetch the given accounts from RPC.
-    /// NOTE: if we get an RPC error we just log it and give up since there is no
-    ///       obvious way how to handle this even if we were to bubble the error up.
-    /// Any action that depends on those accounts to be there will fail.
-    /// NOTE: this is not used during subscription updates since we receive the data
-    ///       as part of that update, thus we won't have stale data issues.
-    #[allow(clippy::too_many_arguments)]
-    fn fetch(
-        &self,
-        pubkeys: Vec<Pubkey>,
-        generations: HashMap<Pubkey, FetchingAccountGeneration>,
-        mark_empty_if_not_found: Option<&[Pubkey]>,
-        min_context_slot: u64,
-        fetch_context: AccountFetchContext,
-    ) {
-        let rpc_client = self.rpc_client.clone();
-        let fetching_accounts = self.fetching_accounts.clone();
-        let commitment = self.rpc_client.commitment();
-        let mark_empty_if_not_found =
-            mark_empty_if_not_found.unwrap_or(&[]).to_vec();
-        let subscription_tiers = self.subscription_tier_ctx();
-        tokio::spawn(async move {
-            use RemoteAccount::*;
-
-            let fetch_started_at = std::time::Instant::now();
-            // Helper to notify all pending requests of fetch failure
-            let notify_error = |error_msg: &str| {
-                let mut fetching = fetching_accounts
-                    .lock()
-                    .unwrap_or_else(|poison| poison.into_inner());
-                warn!(
-                    pubkey_count = pubkeys.len(),
-                    pubkeys = %pubkeys_str(&pubkeys),
-                    min_context_slot = min_context_slot,
-                    commitment = ?commitment,
-                    fetch_entrypoint = %fetch_context.entrypoint(),
-                    fetch_reason = %fetch_context.reason(),
-                    elapsed_ms = fetch_started_at.elapsed().as_millis() as u64,
-                    error = %error_msg,
-                    "{error_msg}"
-                );
-                inc_account_fetches_failed(pubkeys.len() as u64);
-
-                for pubkey in &pubkeys {
-                    // Update metrics
-                    // Remove pending requests and send error
-                    if let Some(generation) = generations.get(pubkey).copied() {
-                        if let Some(state) =
-                            remove_fetching_account_if_generation_matches(
-                                &mut fetching,
-                                pubkey,
-                                generation,
-                            )
-                        {
-                            observe_chainlink_pending_fetch_owner_duration_seconds_with_context(
-                                state.fetch_context,
-                                ChainlinkPendingFetchLayer::RemoteAccountProvider,
-                                ChainlinkPendingFetchOutcome::OwnerFailed,
-                                state.owner_started_at.elapsed().as_secs_f64(),
-                            );
-                            for sender in state.waiters {
-                                let error = RemoteAccountProviderError::AccountResolutionsFailed(
-                                    format!("{}: {}", pubkey, error_msg)
-                                );
-                                let _ = sender.send(Err(error));
-                            }
-                        }
-                    }
-                }
-            };
-
-            let mut remaining_retries: u64 = RPC_FETCH_MAX_RETRIES;
-
-            if tracing::enabled!(tracing::Level::TRACE) {
-                trace!(pubkeys = pubkeys_str(&pubkeys), "Fetching accounts");
-            }
-
-            macro_rules! retry {
-                ($msg:expr) => {{
-                    trace!($msg);
-                    remaining_retries -= 1;
-                    if remaining_retries <= 0 {
-                        let err_msg = format!("Max retries {RPC_FETCH_MAX_RETRIES} reached, giving up on fetching accounts: {pubkeys:?}");
-                        notify_error(&err_msg);
-                        return;
-                    }
-                    tokio::time::sleep(RPC_FETCH_RETRY_DELAY).await;
-                    continue;
-                }};
-            }
-            let response = loop {
-                // We provide the min_context slot in order to _force_ the RPC to update
-                // its account cache. Otherwise we could just keep fetching the accounts
-                // until the context slot is high enough.
-                metrics::inc_remote_account_provider_a_count();
-                match tokio::time::timeout(RPC_FETCH_TIMEOUT, async {
-                    let config = RpcAccountInfoConfig {
-                        commitment: Some(commitment),
-                        min_context_slot: Some(min_context_slot),
-                        encoding: Some(UiAccountEncoding::Base64Zstd),
-                        data_slice: None,
-                    };
-
-                    if pubkeys.len() == 1 {
-                        rpc_client
-                            .get_account_with_config(&pubkeys[0], config)
-                            .await
-                            .map(|res| (res.context.slot, vec![res.value]))
-                    } else {
-                        rpc_client
-                            .get_multiple_accounts_with_config(&pubkeys, config)
-                            .await
-                            .map(|res| (res.context.slot, res.value))
-                    }
-                })
-                .await
-                {
-                    Ok(Ok(res)) => {
-                        let (slot, value) = res;
-                        if slot < min_context_slot {
-                            retry!("Response slot {slot} < {min_context_slot}. Retrying...");
-                        } else {
-                            break (slot, value);
-                        }
-                    }
-                    Ok(Err(err)) => match *err.kind {
-                        ErrorKind::RpcError(rpc_err) => {
-                            match rpc_err {
-                                RpcError::ForUser(ref rpc_user_err) => {
-                                    // When an account is not present for the desired
-                                    // min-context slot then we normally get the below
-                                    // handled `RpcResponseError`, but may also get the
-                                    // following error from the RPC.
-                                    // See test::ixtest_existing_account_for_future_slot
-                                    // ```
-                                    // RpcError(
-                                    //   ForUser(
-                                    //       "AccountNotFound: \
-                                    // pubkey=DaeruQ4SukTQaJA5muyv51MQZok7oaCAF8fAW19mbJv5: \
-                                    //        RPC response error -32016: \
-                                    //        Minimum context slot has not been reached; ",
-                                    //   ),
-                                    // )
-                                    // ```
-                                    retry!("Fetching accounts failed: {rpc_user_err:?}");
-                                 }
-                                RpcError::RpcResponseError {
-                                    code,
-                                    message,
-                                    data,
-                                } => {
-                                    if code == JSON_RPC_SERVER_ERROR_MIN_CONTEXT_SLOT_NOT_REACHED || code == HELIUS_CONTEXT_SLOT_NOT_REACHED {
-                                        retry!("Minimum context slot {min_context_slot} not reached for {commitment:?}. code={code}, message={message}, data={data:?}");
-                                    } else {
-                                        let err = RpcError::RpcResponseError {
-                                            code,
-                                            message,
-                                            data,
-                                        };
-                                        let err_msg = format!(
-                                            "RpcError fetching accounts {}: {err:?}", pubkeys_str(&pubkeys)
-                                        );
-                                        notify_error(&err_msg);
-                                        return;
-                                    }
-                                }
-                                err => {
-                                    let err_msg = format!(
-                                        "RpcError fetching accounts {}: {err:?}", pubkeys_str(&pubkeys)
-                                    );
-                                     notify_error(&err_msg);
-                                     return;
-                                 }
-                            }
-                        }
-                        ErrorKind::Custom(message)
-                            if message
-                                .to_ascii_lowercase()
-                                .contains("minimum context slot") =>
-                        {
-                            retry!(
-                                "Minimum context slot {min_context_slot} not reached for {commitment:?}: {message}"
-                            );
-                        }
-                        _ => {
-                            let err_msg = format!(
-                                "RpcError fetching accounts {}: {err:?}",
-                                pubkeys_str(&pubkeys)
-                            );
-                            notify_error(&err_msg);
-                            return;
-                        }
-                    },
-                    Err(_) => {
-                        let attempt =
-                            RPC_FETCH_MAX_RETRIES - remaining_retries + 1;
-                        warn!(
-                                pubkey_count = pubkeys.len(),
-                                pubkeys = %pubkeys_str(&pubkeys),
-                                attempt = attempt,
-                                max_retries = RPC_FETCH_MAX_RETRIES,
-                                remaining_retries = remaining_retries.saturating_sub(1),
-                                timeout_ms = RPC_FETCH_TIMEOUT.as_millis() as u64,
-                                elapsed_ms = fetch_started_at.elapsed().as_millis() as u64,
-                                min_context_slot = min_context_slot,
-                                commitment = ?commitment,
-                                fetch_entrypoint = %fetch_context.entrypoint(),
-                        fetch_reason = %fetch_context.reason(),
-                                "RPC call timeout. Retrying..."
-                            );
-                        remaining_retries -= 1;
-                        if remaining_retries == 0 {
-                            let err_msg = format!("Max retries {RPC_FETCH_MAX_RETRIES} reached, giving up on fetching accounts: {pubkeys:?}");
-                            notify_error(&err_msg);
-                            return;
-                        }
-                        tokio::time::sleep(RPC_FETCH_RETRY_DELAY).await;
-                        continue;
-                    }
-                };
-            };
-
-            // TODO: should we retry if not or respond with an error?
-            let (response_slot, response_value) = response;
-            assert!(response_slot >= min_context_slot);
-
-            if response_value.len() != pubkeys.len() {
-                let err_msg = format!(
-                    "RPC returned {} account results for {} requested accounts: {}",
-                    response_value.len(),
-                    pubkeys.len(),
-                    pubkeys_str(&pubkeys)
-                );
-                notify_error(&err_msg);
-                return;
-            }
-
-            let mut found_count = 0u64;
-            let mut not_found_count = 0u64;
-            let mut not_found_pubkeys = HashSet::new();
-
-            let remote_accounts: Vec<RemoteAccount> = pubkeys
-                .iter()
-                .zip(response_value)
-                .map(|(pubkey, acc)| match acc {
-                    Some(value) => {
-                        found_count += 1;
-                        RemoteAccount::from_fresh_account(
-                            value,
-                            response_slot,
-                            RemoteAccountUpdateSource::Fetch,
-                        )
-                    }
-                    None if mark_empty_if_not_found.contains(pubkey) => {
-                        not_found_count += 1;
-                        not_found_pubkeys.insert(*pubkey);
-                        inc_chainlink_empty_placeholder_accounts_total_with_context(
-                            fetch_context.clone(),
-                            ChainlinkEmptyPlaceholderStage::ConvertedToEmpty,
-                            Outcome::Success,
-                        );
-                        RemoteAccount::from_fresh_account(
-                            Account {
-                                lamports: 0,
-                                data: vec![],
-                                owner: Pubkey::default(),
-                                executable: false,
-                                rent_epoch: 0,
-                            },
-                            response_slot,
-                            RemoteAccountUpdateSource::Fetch,
-                        )
-                    }
-                    None => {
-                        not_found_count += 1;
-                        not_found_pubkeys.insert(*pubkey);
-                        NotFound(response_slot)
-                    }
-                })
-                .collect();
-
-            // Update metrics for successful RPC fetch
-            inc_account_fetches_success(pubkeys.len() as u64);
-            inc_account_fetches_found_with_context(
-                fetch_context.clone(),
-                found_count,
-            );
-            inc_account_fetches_not_found_with_context(
-                fetch_context.clone(),
-                not_found_count,
-            );
-
-            if tracing::enabled!(tracing::Level::TRACE) {
-                let pubkeys = pubkeys
-                    .iter()
-                    .map(|pk| pk.to_string())
-                    .collect::<Vec<_>>()
-                    .join(", ");
-                trace!(
-                    pubkeys = %pubkeys, remote_accounts = ?remote_accounts, "Fetched, notifying pending requests"
-                );
-            }
-
-            // Notify all pending requests with fetch results (unless subscription override occurred)
-            for (pubkey, remote_account) in
-                pubkeys.iter().zip(remote_accounts.iter())
-            {
-                let (waiters, classification_result) = {
-                    // The per-key guard serializes this resolution against
-                    // subscription updates and other transitions of the same
-                    // key; the tier helpers scope the transition lock to
-                    // their in-memory critical sections.
-                    let _subscription_guard =
-                        subscription_key_owned_guard_from_map(
-                            &subscription_tiers.subscription_key_locks,
-                            *pubkey,
-                        )
-                        .await;
-                    // Remove from fetching and get pending requests
-                    // Note: the account might have been resolved by a
-                    // subscription update already or replaced by a newer owner.
-                    let Some(generation) = generations.get(pubkey).copied()
-                    else {
-                        continue;
-                    };
-                    let state = {
-                        let mut fetching = fetching_accounts
-                            .lock()
-                            .unwrap_or_else(|poison| poison.into_inner());
-                        remove_fetching_account_if_generation_matches(
-                            &mut fetching,
-                            pubkey,
-                            generation,
-                        )
-                    };
-                    if let Some(state) = state {
-                        let waiters = state.waiters;
-
-                        let classification_result = subscription_tiers
-                            .apply_fetch_classification(
-                                pubkey,
-                                response_slot,
-                                not_found_pubkeys.contains(pubkey),
-                            )
-                            .await;
-                        observe_chainlink_pending_fetch_owner_duration_seconds_with_context(
-                            state.fetch_context.clone(),
-                            ChainlinkPendingFetchLayer::RemoteAccountProvider,
-                            if classification_result.is_ok() {
-                                ChainlinkPendingFetchOutcome::OwnerSucceeded
-                            } else {
-                                ChainlinkPendingFetchOutcome::OwnerFailed
-                            },
-                            state.owner_started_at.elapsed().as_secs_f64(),
-                        );
-                        (waiters, classification_result)
-                    } else {
-                        inc_chainlink_pending_fetch_accounts_with_context(
-                            fetch_context.clone(),
-                            ChainlinkPendingFetchLayer::RemoteAccountProvider,
-                            ChainlinkPendingFetchOutcome::RpcFetchCompletedAfterUpdate,
-                            1,
-                        );
-                        if tracing::enabled!(tracing::Level::TRACE) {
-                            trace!(
-                                "Account {pubkey} generation {generation} was already resolved or replaced"
-                            );
-                        }
-                        continue;
-                    }
-                };
-
-                // Send the fetch result to all waiting requests
-                for request in waiters {
-                    let result = match &classification_result {
-                        Ok(()) => Ok(remote_account.clone()),
-                        Err(RemoteAccountProviderError::NoEvictableSubscriptionCapacity { pubkey }) => {
-                            Err(RemoteAccountProviderError::NoEvictableSubscriptionCapacity {
-                                pubkey: *pubkey,
-                            })
-                        }
-                        Err(err) => Err(
-                            RemoteAccountProviderError::AccountResolutionsFailed(
-                                err.to_string(),
-                            ),
-                        ),
-                    };
-                    let _ = request.send(result);
-                }
-            }
-        });
     }
 }
